python_legacy/pyproject.toml
[tool.poetry]
name = "sqlbucks"
version = "0.1.0"
description = ""
authors = ["Aishwarya Setia <setiaaishwarya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.1.0"
bcrypt = "^4.2.0"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
